minor bug fixes

+++ src/var/String.cpp
@@ -184,7 +184,7 @@ Vector<String> String::split(const String & delimiter) const {
 
 
 size_t String::count(const var::String & occurance) const {
-	u32 pos = 0;
+	size_t pos = 0;
 	u32 item_count = 0;
 	while( pos < length() && pos != String::npos){
 		pos = find(occurance, String::Position(pos));
